Remove redundant factory code
