+++ src/main/java/com/minecolonies/coremod/colony/requestsystem/init/StandardFactoryControllerInitializer.java
@@ -7,7 +7,6 @@ import com.minecolonies.api.colony.requestsystem.token.InitializedTokenFactory;
 import com.minecolonies.api.colony.requestsystem.token.RandomSeededTokenFactory;
 import com.minecolonies.api.colony.requestsystem.token.StandardTokenFactory;
 import com.minecolonies.coremod.colony.buildings.modules.settings.SettingsFactories;
-import com.minecolonies.coremod.colony.crafting.CustomRecipeFactory;
 import com.minecolonies.coremod.colony.crafting.ImmutableItemStorageFactory;
 import com.minecolonies.coremod.colony.crafting.ItemStorageFactory;
 import com.minecolonies.coremod.colony.crafting.RecipeStorageFactory;
@@ -67,7 +66,6 @@ public final class StandardFactoryControllerInitializer
         StandardFactoryController.getInstance().registerNewFactory(new ImmutableItemStorageFactory());
         StandardFactoryController.getInstance().registerNewFactory(new GlobalResearchFactory());
         StandardFactoryController.getInstance().registerNewFactory(new LocalResearchFactory());
-        StandardFactoryController.getInstance().registerNewFactory(new CustomRecipeFactory());
         StandardFactoryController.getInstance().registerNewFactory(new IntegerFactory());
         StandardFactoryController.getInstance().registerNewFactory(new TypeTokenFactory());
         StandardFactoryController.getInstance().registerNewFactory(new StandardRequestIdentitiesDataStore.Factory());
